requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-environ==0.10.0
psycopg2==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.3
